requirements.txt
# 1. Document Processing
pymupdf==1.24.9

# 2. Chunking & Orchestration
langchain==0.2.12

# 3. Embeddings + LLM
openai==1.42.0

# 4. Vector Database
faiss-cpu==1.12.0

# 5. UI Layer
streamlit==1.38.0

sentence-transformers==2.2.2   # ✅ added to replace openai embeddings for no cost

huggingface-hub==0.24.6